修正User.Id、Chat.Id的类型为int64

+++ types.go
@@ -1,7 +1,7 @@
 package telegram
 
 type User struct {
-	Id                      int
+	Id                      int64
 	IsBot                   bool   `json:"is_bot"`
 	FirstName               string `json:"first_name"`
 	LastName                string `json:"last_name"`
@@ -12,7 +12,7 @@ type User struct {
 	SupportsInlineQueries   bool   `json:"supports_inline_queries"`
 }
 type Chat struct {
-	Id                    int              `json:"id"`                       //Unique identifier for this chat. This number may have more than 32 significant bits and some programming languages may have difficulty/silent defects in interpreting it. But it has at most 52 significant bits, so a signed 64-bit integer or double-precision float type are safe for storing this identifier.
+	Id                    int64            `json:"id"`                       //Unique identifier for this chat. This number may have more than 32 significant bits and some programming languages may have difficulty/silent defects in interpreting it. But it has at most 52 significant bits, so a signed 64-bit integer or double-precision float type are safe for storing this identifier.
 	Type                  string           `json:"type"`                     //Type of chat, can be either “private”, “group”, “supergroup” or “channel”
 	Title                 string           `json:"title"`                    //Optional. Title, for supergroups, channels and group chats
 	Username              string           `json:"username"`                 //Optional. Username, for private chats, supergroups and channels if available
